backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.10.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-cors==0.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
fake-useragent==2.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instaloader==4.14.1
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.50.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pyee==12.1.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
TikTokApi==7.0.0
tk==0.1.0
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.4.7
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
